+++ guides/react-apollo/src/lib/graphql/__generated__/graphql.ts
@@ -32,6 +32,7 @@ export type Scalars = {
   bytea: { input: Buffer; output: Buffer };
   citext: { input: string; output: string };
   jsonb: { input: Record<string, any>; output: Record<string, any> };
+  numeric: { input: any; output: any };
   timestamptz: { input: string; output: string };
   uuid: { input: string; output: string };
 };
@@ -3153,6 +3154,283 @@ export type Jsonb_Comparison_Exp = {
   _nin?: InputMaybe<Array<Scalars["jsonb"]["input"]>>;
 };
 
+/** columns and relationships of "movies" */
+export type Movies = {
+  __typename?: "movies";
+  created_at?: Maybe<Scalars["timestamptz"]["output"]>;
+  director?: Maybe<Scalars["String"]["output"]>;
+  genre?: Maybe<Scalars["String"]["output"]>;
+  id: Scalars["uuid"]["output"];
+  rating?: Maybe<Scalars["numeric"]["output"]>;
+  release_year?: Maybe<Scalars["Int"]["output"]>;
+  title: Scalars["String"]["output"];
+  updated_at?: Maybe<Scalars["timestamptz"]["output"]>;
+};
+
+/** aggregated selection of "movies" */
+export type Movies_Aggregate = {
+  __typename?: "movies_aggregate";
+  aggregate?: Maybe<Movies_Aggregate_Fields>;
+  nodes: Array<Movies>;
+};
+
+/** aggregate fields of "movies" */
+export type Movies_Aggregate_Fields = {
+  __typename?: "movies_aggregate_fields";
+  avg?: Maybe<Movies_Avg_Fields>;
+  count: Scalars["Int"]["output"];
+  max?: Maybe<Movies_Max_Fields>;
+  min?: Maybe<Movies_Min_Fields>;
+  stddev?: Maybe<Movies_Stddev_Fields>;
+  stddev_pop?: Maybe<Movies_Stddev_Pop_Fields>;
+  stddev_samp?: Maybe<Movies_Stddev_Samp_Fields>;
+  sum?: Maybe<Movies_Sum_Fields>;
+  var_pop?: Maybe<Movies_Var_Pop_Fields>;
+  var_samp?: Maybe<Movies_Var_Samp_Fields>;
+  variance?: Maybe<Movies_Variance_Fields>;
+};
+
+/** aggregate fields of "movies" */
+export type Movies_Aggregate_FieldsCountArgs = {
+  columns?: InputMaybe<Array<Movies_Select_Column>>;
+  distinct?: InputMaybe<Scalars["Boolean"]["input"]>;
+};
+
+/** aggregate avg on columns */
+export type Movies_Avg_Fields = {
+  __typename?: "movies_avg_fields";
+  rating?: Maybe<Scalars["Float"]["output"]>;
+  release_year?: Maybe<Scalars["Float"]["output"]>;
+};
+
+/** Boolean expression to filter rows from the table "movies". All fields are combined with a logical 'AND'. */
+export type Movies_Bool_Exp = {
+  _and?: InputMaybe<Array<Movies_Bool_Exp>>;
+  _not?: InputMaybe<Movies_Bool_Exp>;
+  _or?: InputMaybe<Array<Movies_Bool_Exp>>;
+  created_at?: InputMaybe<Timestamptz_Comparison_Exp>;
+  director?: InputMaybe<String_Comparison_Exp>;
+  genre?: InputMaybe<String_Comparison_Exp>;
+  id?: InputMaybe<Uuid_Comparison_Exp>;
+  rating?: InputMaybe<Numeric_Comparison_Exp>;
+  release_year?: InputMaybe<Int_Comparison_Exp>;
+  title?: InputMaybe<String_Comparison_Exp>;
+  updated_at?: InputMaybe<Timestamptz_Comparison_Exp>;
+};
+
+/** unique or primary key constraints on table "movies" */
+export enum Movies_Constraint {
+  /** unique or primary key constraint on columns "id" */
+  MoviesPkey = "movies_pkey",
+}
+
+/** input type for incrementing numeric columns in table "movies" */
+export type Movies_Inc_Input = {
+  rating?: InputMaybe<Scalars["numeric"]["input"]>;
+  release_year?: InputMaybe<Scalars["Int"]["input"]>;
+};
+
+/** input type for inserting data into table "movies" */
+export type Movies_Insert_Input = {
+  created_at?: InputMaybe<Scalars["timestamptz"]["input"]>;
+  director?: InputMaybe<Scalars["String"]["input"]>;
+  genre?: InputMaybe<Scalars["String"]["input"]>;
+  id?: InputMaybe<Scalars["uuid"]["input"]>;
+  rating?: InputMaybe<Scalars["numeric"]["input"]>;
+  release_year?: InputMaybe<Scalars["Int"]["input"]>;
+  title?: InputMaybe<Scalars["String"]["input"]>;
+  updated_at?: InputMaybe<Scalars["timestamptz"]["input"]>;
+};
+
+/** aggregate max on columns */
+export type Movies_Max_Fields = {
+  __typename?: "movies_max_fields";
+  created_at?: Maybe<Scalars["timestamptz"]["output"]>;
+  director?: Maybe<Scalars["String"]["output"]>;
+  genre?: Maybe<Scalars["String"]["output"]>;
+  id?: Maybe<Scalars["uuid"]["output"]>;
+  rating?: Maybe<Scalars["numeric"]["output"]>;
+  release_year?: Maybe<Scalars["Int"]["output"]>;
+  title?: Maybe<Scalars["String"]["output"]>;
+  updated_at?: Maybe<Scalars["timestamptz"]["output"]>;
+};
+
+/** aggregate min on columns */
+export type Movies_Min_Fields = {
+  __typename?: "movies_min_fields";
+  created_at?: Maybe<Scalars["timestamptz"]["output"]>;
+  director?: Maybe<Scalars["String"]["output"]>;
+  genre?: Maybe<Scalars["String"]["output"]>;
+  id?: Maybe<Scalars["uuid"]["output"]>;
+  rating?: Maybe<Scalars["numeric"]["output"]>;
+  release_year?: Maybe<Scalars["Int"]["output"]>;
+  title?: Maybe<Scalars["String"]["output"]>;
+  updated_at?: Maybe<Scalars["timestamptz"]["output"]>;
+};
+
+/** response of any mutation on the table "movies" */
+export type Movies_Mutation_Response = {
+  __typename?: "movies_mutation_response";
+  /** number of rows affected by the mutation */
+  affected_rows: Scalars["Int"]["output"];
+  /** data from the rows affected by the mutation */
+  returning: Array<Movies>;
+};
+
+/** on_conflict condition type for table "movies" */
+export type Movies_On_Conflict = {
+  constraint: Movies_Constraint;
+  update_columns?: Array<Movies_Update_Column>;
+  where?: InputMaybe<Movies_Bool_Exp>;
+};
+
+/** Ordering options when selecting data from "movies". */
+export type Movies_Order_By = {
+  created_at?: InputMaybe<Order_By>;
+  director?: InputMaybe<Order_By>;
+  genre?: InputMaybe<Order_By>;
+  id?: InputMaybe<Order_By>;
+  rating?: InputMaybe<Order_By>;
+  release_year?: InputMaybe<Order_By>;
+  title?: InputMaybe<Order_By>;
+  updated_at?: InputMaybe<Order_By>;
+};
+
+/** primary key columns input for table: movies */
+export type Movies_Pk_Columns_Input = {
+  id: Scalars["uuid"]["input"];
+};
+
+/** select columns of table "movies" */
+export enum Movies_Select_Column {
+  /** column name */
+  CreatedAt = "created_at",
+  /** column name */
+  Director = "director",
+  /** column name */
+  Genre = "genre",
+  /** column name */
+  Id = "id",
+  /** column name */
+  Rating = "rating",
+  /** column name */
+  ReleaseYear = "release_year",
+  /** column name */
+  Title = "title",
+  /** column name */
+  UpdatedAt = "updated_at",
+}
+
+/** input type for updating data in table "movies" */
+export type Movies_Set_Input = {
+  created_at?: InputMaybe<Scalars["timestamptz"]["input"]>;
+  director?: InputMaybe<Scalars["String"]["input"]>;
+  genre?: InputMaybe<Scalars["String"]["input"]>;
+  id?: InputMaybe<Scalars["uuid"]["input"]>;
+  rating?: InputMaybe<Scalars["numeric"]["input"]>;
+  release_year?: InputMaybe<Scalars["Int"]["input"]>;
+  title?: InputMaybe<Scalars["String"]["input"]>;
+  updated_at?: InputMaybe<Scalars["timestamptz"]["input"]>;
+};
+
+/** aggregate stddev on columns */
+export type Movies_Stddev_Fields = {
+  __typename?: "movies_stddev_fields";
+  rating?: Maybe<Scalars["Float"]["output"]>;
+  release_year?: Maybe<Scalars["Float"]["output"]>;
+};
+
+/** aggregate stddev_pop on columns */
+export type Movies_Stddev_Pop_Fields = {
+  __typename?: "movies_stddev_pop_fields";
+  rating?: Maybe<Scalars["Float"]["output"]>;
+  release_year?: Maybe<Scalars["Float"]["output"]>;
+};
+
+/** aggregate stddev_samp on columns */
+export type Movies_Stddev_Samp_Fields = {
+  __typename?: "movies_stddev_samp_fields";
+  rating?: Maybe<Scalars["Float"]["output"]>;
+  release_year?: Maybe<Scalars["Float"]["output"]>;
+};
+
+/** Streaming cursor of the table "movies" */
+export type Movies_Stream_Cursor_Input = {
+  /** Stream column input with initial value */
+  initial_value: Movies_Stream_Cursor_Value_Input;
+  /** cursor ordering */
+  ordering?: InputMaybe<Cursor_Ordering>;
+};
+
+/** Initial value of the column from where the streaming should start */
+export type Movies_Stream_Cursor_Value_Input = {
+  created_at?: InputMaybe<Scalars["timestamptz"]["input"]>;
+  director?: InputMaybe<Scalars["String"]["input"]>;
+  genre?: InputMaybe<Scalars["String"]["input"]>;
+  id?: InputMaybe<Scalars["uuid"]["input"]>;
+  rating?: InputMaybe<Scalars["numeric"]["input"]>;
+  release_year?: InputMaybe<Scalars["Int"]["input"]>;
+  title?: InputMaybe<Scalars["String"]["input"]>;
+  updated_at?: InputMaybe<Scalars["timestamptz"]["input"]>;
+};
+
+/** aggregate sum on columns */
+export type Movies_Sum_Fields = {
+  __typename?: "movies_sum_fields";
+  rating?: Maybe<Scalars["numeric"]["output"]>;
+  release_year?: Maybe<Scalars["Int"]["output"]>;
+};
+
+/** update columns of table "movies" */
+export enum Movies_Update_Column {
+  /** column name */
+  CreatedAt = "created_at",
+  /** column name */
+  Director = "director",
+  /** column name */
+  Genre = "genre",
+  /** column name */
+  Id = "id",
+  /** column name */
+  Rating = "rating",
+  /** column name */
+  ReleaseYear = "release_year",
+  /** column name */
+  Title = "title",
+  /** column name */
+  UpdatedAt = "updated_at",
+}
+
+export type Movies_Updates = {
+  /** increments the numeric columns with given value of the filtered values */
+  _inc?: InputMaybe<Movies_Inc_Input>;
+  /** sets the columns of the filtered rows to the given values */
+  _set?: InputMaybe<Movies_Set_Input>;
+  /** filter the rows which have to be updated */
+  where: Movies_Bool_Exp;
+};
+
+/** aggregate var_pop on columns */
+export type Movies_Var_Pop_Fields = {
+  __typename?: "movies_var_pop_fields";
+  rating?: Maybe<Scalars["Float"]["output"]>;
+  release_year?: Maybe<Scalars["Float"]["output"]>;
+};
+
+/** aggregate var_samp on columns */
+export type Movies_Var_Samp_Fields = {
+  __typename?: "movies_var_samp_fields";
+  rating?: Maybe<Scalars["Float"]["output"]>;
+  release_year?: Maybe<Scalars["Float"]["output"]>;
+};
+
+/** aggregate variance on columns */
+export type Movies_Variance_Fields = {
+  __typename?: "movies_variance_fields";
+  rating?: Maybe<Scalars["Float"]["output"]>;
+  release_year?: Maybe<Scalars["Float"]["output"]>;
+};
+
 /** mutation root */
 export type Mutation_Root = {
   __typename?: "mutation_root";
@@ -3220,6 +3498,10 @@ export type Mutation_Root = {
   deleteVirus?: Maybe<Virus>;
   /** delete data from the table: "storage.virus" */
   deleteViruses?: Maybe<Virus_Mutation_Response>;
+  /** delete data from the table: "movies" */
+  delete_movies?: Maybe<Movies_Mutation_Response>;
+  /** delete single row from the table: "movies" */
+  delete_movies_by_pk?: Maybe<Movies>;
   /** insert a single row into the table: "attachments" */
   insertAttachment?: Maybe<Attachments>;
   /** insert data into the table: "attachments" */
@@ -3284,6 +3566,10 @@ export type Mutation_Root = {
   insertVirus?: Maybe<Virus>;
   /** insert data into the table: "storage.virus" */
   insertViruses?: Maybe<Virus_Mutation_Response>;
+  /** insert data into the table: "movies" */
+  insert_movies?: Maybe<Movies_Mutation_Response>;
+  /** insert a single row into the table: "movies" */
+  insert_movies_one?: Maybe<Movies>;
   /** update single row of the table: "attachments" */
   updateAttachment?: Maybe<Attachments>;
   /** update data of the table: "attachments" */
@@ -3390,6 +3676,12 @@ export type Mutation_Root = {
   update_buckets_many?: Maybe<Array<Maybe<Buckets_Mutation_Response>>>;
   /** update multiples rows of table: "storage.files" */
   update_files_many?: Maybe<Array<Maybe<Files_Mutation_Response>>>;
+  /** update data of the table: "movies" */
+  update_movies?: Maybe<Movies_Mutation_Response>;
+  /** update single row of the table: "movies" */
+  update_movies_by_pk?: Maybe<Movies>;
+  /** update multiples rows of table: "movies" */
+  update_movies_many?: Maybe<Array<Maybe<Movies_Mutation_Response>>>;
   /** update multiples rows of table: "auth.users" */
   update_users_many?: Maybe<Array<Maybe<Users_Mutation_Response>>>;
   /** update multiples rows of table: "storage.virus" */
@@ -3555,6 +3847,16 @@ export type Mutation_RootDeleteVirusArgs = {
 /** mutation root */
 export type Mutation_RootDeleteVirusesArgs = {
   where: Virus_Bool_Exp;
+};
+
+/** mutation root */
+export type Mutation_RootDelete_MoviesArgs = {
+  where: Movies_Bool_Exp;
+};
+
+/** mutation root */
+export type Mutation_RootDelete_Movies_By_PkArgs = {
+  id: Scalars["uuid"]["input"];
 };
 
 /** mutation root */
@@ -3747,6 +4049,18 @@ export type Mutation_RootInsertVirusArgs = {
 export type Mutation_RootInsertVirusesArgs = {
   objects: Array<Virus_Insert_Input>;
   on_conflict?: InputMaybe<Virus_On_Conflict>;
+};
+
+/** mutation root */
+export type Mutation_RootInsert_MoviesArgs = {
+  objects: Array<Movies_Insert_Input>;
+  on_conflict?: InputMaybe<Movies_On_Conflict>;
+};
+
+/** mutation root */
+export type Mutation_RootInsert_Movies_OneArgs = {
+  object: Movies_Insert_Input;
+  on_conflict?: InputMaybe<Movies_On_Conflict>;
 };
 
 /** mutation root */
@@ -4068,6 +4382,25 @@ export type Mutation_RootUpdate_Files_ManyArgs = {
 };
 
 /** mutation root */
+export type Mutation_RootUpdate_MoviesArgs = {
+  _inc?: InputMaybe<Movies_Inc_Input>;
+  _set?: InputMaybe<Movies_Set_Input>;
+  where: Movies_Bool_Exp;
+};
+
+/** mutation root */
+export type Mutation_RootUpdate_Movies_By_PkArgs = {
+  _inc?: InputMaybe<Movies_Inc_Input>;
+  _set?: InputMaybe<Movies_Set_Input>;
+  pk_columns: Movies_Pk_Columns_Input;
+};
+
+/** mutation root */
+export type Mutation_RootUpdate_Movies_ManyArgs = {
+  updates: Array<Movies_Updates>;
+};
+
+/** mutation root */
 export type Mutation_RootUpdate_Users_ManyArgs = {
   updates: Array<Users_Updates>;
 };
@@ -4279,6 +4612,19 @@ export type NinjaTurtles_Updates = {
   where: NinjaTurtles_Bool_Exp;
 };
 
+/** Boolean expression to compare columns of type "numeric". All fields are combined with logical 'AND'. */
+export type Numeric_Comparison_Exp = {
+  _eq?: InputMaybe<Scalars["numeric"]["input"]>;
+  _gt?: InputMaybe<Scalars["numeric"]["input"]>;
+  _gte?: InputMaybe<Scalars["numeric"]["input"]>;
+  _in?: InputMaybe<Array<Scalars["numeric"]["input"]>>;
+  _is_null?: InputMaybe<Scalars["Boolean"]["input"]>;
+  _lt?: InputMaybe<Scalars["numeric"]["input"]>;
+  _lte?: InputMaybe<Scalars["numeric"]["input"]>;
+  _neq?: InputMaybe<Scalars["numeric"]["input"]>;
+  _nin?: InputMaybe<Array<Scalars["numeric"]["input"]>>;
+};
+
 /** column ordering options */
 export enum Order_By {
   /** in ascending order, nulls last */
@@ -4369,6 +4715,12 @@ export type Query_Root = {
   files: Array<Files>;
   /** fetch aggregated fields from the table: "storage.files" */
   filesAggregate: Files_Aggregate;
+  /** fetch data from the table: "movies" */
+  movies: Array<Movies>;
+  /** fetch aggregated fields from the table: "movies" */
+  movies_aggregate: Movies_Aggregate;
+  /** fetch data from the table: "movies" using primary key columns */
+  movies_by_pk?: Maybe<Movies>;
   /** fetch data from the table: "ninja_turtles" using primary key columns */
   ninjaTurtle?: Maybe<NinjaTurtles>;
   /** fetch data from the table: "ninja_turtles" */
@@ -4636,6 +4988,26 @@ export type Query_RootFilesAggregateArgs = {
   where?: InputMaybe<Files_Bool_Exp>;
 };
 
+export type Query_RootMoviesArgs = {
+  distinct_on?: InputMaybe<Array<Movies_Select_Column>>;
+  limit?: InputMaybe<Scalars["Int"]["input"]>;
+  offset?: InputMaybe<Scalars["Int"]["input"]>;
+  order_by?: InputMaybe<Array<Movies_Order_By>>;
+  where?: InputMaybe<Movies_Bool_Exp>;
+};
+
+export type Query_RootMovies_AggregateArgs = {
+  distinct_on?: InputMaybe<Array<Movies_Select_Column>>;
+  limit?: InputMaybe<Scalars["Int"]["input"]>;
+  offset?: InputMaybe<Scalars["Int"]["input"]>;
+  order_by?: InputMaybe<Array<Movies_Order_By>>;
+  where?: InputMaybe<Movies_Bool_Exp>;
+};
+
+export type Query_RootMovies_By_PkArgs = {
+  id: Scalars["uuid"]["input"];
+};
+
 export type Query_RootNinjaTurtleArgs = {
   id: Scalars["uuid"]["input"];
 };
@@ -4814,6 +5186,14 @@ export type Subscription_Root = {
   filesAggregate: Files_Aggregate;
   /** fetch data from the table in a streaming manner: "storage.files" */
   files_stream: Array<Files>;
+  /** fetch data from the table: "movies" */
+  movies: Array<Movies>;
+  /** fetch aggregated fields from the table: "movies" */
+  movies_aggregate: Movies_Aggregate;
+  /** fetch data from the table: "movies" using primary key columns */
+  movies_by_pk?: Maybe<Movies>;
+  /** fetch data from the table in a streaming manner: "movies" */
+  movies_stream: Array<Movies>;
   /** fetch data from the table: "ninja_turtles" using primary key columns */
   ninjaTurtle?: Maybe<NinjaTurtles>;
   /** fetch data from the table: "ninja_turtles" */
@@ -5159,6 +5539,32 @@ export type Subscription_RootFiles_StreamArgs = {
   batch_size: Scalars["Int"]["input"];
   cursor: Array<InputMaybe<Files_Stream_Cursor_Input>>;
   where?: InputMaybe<Files_Bool_Exp>;
+};
+
+export type Subscription_RootMoviesArgs = {
+  distinct_on?: InputMaybe<Array<Movies_Select_Column>>;
+  limit?: InputMaybe<Scalars["Int"]["input"]>;
+  offset?: InputMaybe<Scalars["Int"]["input"]>;
+  order_by?: InputMaybe<Array<Movies_Order_By>>;
+  where?: InputMaybe<Movies_Bool_Exp>;
+};
+
+export type Subscription_RootMovies_AggregateArgs = {
+  distinct_on?: InputMaybe<Array<Movies_Select_Column>>;
+  limit?: InputMaybe<Scalars["Int"]["input"]>;
+  offset?: InputMaybe<Scalars["Int"]["input"]>;
+  order_by?: InputMaybe<Array<Movies_Order_By>>;
+  where?: InputMaybe<Movies_Bool_Exp>;
+};
+
+export type Subscription_RootMovies_By_PkArgs = {
+  id: Scalars["uuid"]["input"];
+};
+
+export type Subscription_RootMovies_StreamArgs = {
+  batch_size: Scalars["Int"]["input"];
+  cursor: Array<InputMaybe<Movies_Stream_Cursor_Input>>;
+  where?: InputMaybe<Movies_Bool_Exp>;
 };
 
 export type Subscription_RootNinjaTurtleArgs = {