+++ packages/nhost-js/src/auth/__tests__/docstrings.test.ts
@@ -1,29 +1,34 @@
 import { test, expect } from "@jest/globals";
-import { createClient } from "../../";
-import { type ErrorResponse } from "../../auth";
-import { type FetchError } from "../../fetch";
+import { createClient } from "@nhost/nhost-js";
+import { FetchError } from "@nhost/nhost-js/fetch";
 
-test("error handling for auth", async () => {
-  const subdomain = "local";
-  const region = "local";
+const subdomain = "local";
+const region = "local";
 
+test("usage", async () => {
   const email = `test-${Date.now()}@example.com`;
   const password = "password123";
 
-  //#region errorHandling
-  // Needs the following imports:
-  //
-  // import {
-  //   type ErrorResponse,
-  // } from "@nhost/nhost-js/auth";
-  // import {
-  //   type FetchError,
-  // } from "@nhost/nhost-js/fetch";
-  //
   const nhost = createClient({
     subdomain,
     region,
   });
+
+  await nhost.auth.signUpEmailPassword({
+    email,
+    password,
+  });
+});
+
+test("error handling for auth", async () => {
+  const email = `test-${Date.now()}@example.com`;
+  const password = "password123";
+
+  const nhost = createClient({
+    subdomain,
+    region,
+  });
+
   try {
     await nhost.auth.signInEmailPassword({
       email,
@@ -31,8 +36,11 @@ test("error handling for auth", async () => {
     });
 
     expect(true).toBe(false); // This should not be reached
-  } catch (error) {
-    const err = error as FetchError<ErrorResponse>;
+  } catch (err) {
+    if (!(err instanceof FetchError)) {
+      throw err; // Re-throw if it's not a FetchError
+    }
+
     console.log("Error:", err);
     // Error: {
     //   body: {
@@ -50,12 +58,39 @@ test("error handling for auth", async () => {
 
     // error handling...
 
-    //#endregion errorHandling
     expect(err.status).toBe(401);
     expect(err.body).toStrictEqual({
       error: "invalid-email-password",
       message: "Incorrect email or password",
       status: 401,
     });
+  }
+});
+
+test("error handling for auth error type", async () => {
+  const email = `test-${Date.now()}@example.com`;
+  const password = "password123";
+
+  const nhost = createClient({
+    subdomain,
+    region,
+  });
+
+  try {
+    await nhost.auth.signInEmailPassword({
+      email,
+      password,
+    });
+
+    expect(true).toBe(false); // This should not be reached
+  } catch (err) {
+    if (!(err instanceof Error)) {
+      throw err; // Re-throw if it's not an Error
+    }
+
+    console.log("Error:", err.message);
+    // Error: Incorrect email or password
+
+    expect(err.message).toBe("Incorrect email or password");
   }
 });