+++ tools/codegen/CHANGELOG.md
@@ -1,2 +1,3 @@
+- 20250704 feat (packages/nhost-js): added functions.post method and improve documentation #43
 - 20250610 fix (tools/codegen;packages/nhost-js): encode URL parameters better #30
 - 20250601 chore (ci): enabled #26